+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file. For commit 
  **Notes**:
  - build requirement change -- from node 18 > node 20 with npm 9.0.0+
  - version is not shown for unauthenticated users #1444
+ - fixed signup login #1444
  - update makefile linker flags to properly set version and sha #1474
  - better windows build support -- `make setup` and `make dev` work as long as git is installed on windows
  - added better first initilization detection